+++ app/src/main/java/com/example/scarlet/feature_training_log/domain/model/Session.kt
@@ -4,8 +4,8 @@ import java.io.Serializable
 import java.util.Date
 
 data class Session(
-    val id: Int = 0,
-    val blockId: Int = 0,
+    val id: Long = 0,
+    val blockId: Long = 0,
     val date: Date = Date(System.currentTimeMillis()),
 ): Serializable
