+++ app/src/main/java/com/example/scarlet/feature_training_log/domain/use_case/session/InsertSetUseCase.kt
@@ -17,7 +17,7 @@ class InsertSetUseCase(
      * @return a resource with data (id of the inserted set)
      */
     suspend operator fun invoke(
-        exerciseId: Int,
+        exerciseId: Long,
         exerciseSets: List<Set>
     ): Resource<Long> {
         val set = Set(
