+++ app/src/main/java/com/example/scarlet/feature_training_log/presentation/block/BlockViewModel.kt
@@ -47,7 +47,7 @@ class BlockViewModel @Inject constructor(
                             resource.data?.let { insertedSessionId ->
                                 _uiActions.emit(UiAction.NavigateToSessionScreen(
                                     session.copy(
-                                        id = insertedSessionId.toInt()
+                                        id = insertedSessionId
                                     )
                                 ))
                             }
